MLK-14938-9 imx8: bootaux: Add i.MX8 M4 boot support
1. Adjust bootaux interface to adapt 64bits address and add core id.
2. Implement bootaux for the M4 boot on i.MX8QM and QXP. Users need to download
   M4 image to any DDR address first. Then use the
   "bootaux <M4 download DDR address> [M4 core id]" to boot CM4_0
   or CM4_1, the default core id is 0 for CM4_0.

   Since current M4 only supports running in TCM. The bootaux will copy
   the M4 image from DDR to its TCML.

3. Implment bootaux for HIFI on QXP
   command: bootaux 0x81000000 1

+++ arch/arm/cpu/armv7/mx6/soc.c
@@ -899,7 +899,7 @@ void imx_setup_hdmi(void)
 #endif
 
 #ifdef CONFIG_IMX_BOOTAUX
-int arch_auxiliary_core_up(u32 core_id, u32 boot_private_data)
+int arch_auxiliary_core_up(u32 core_id, ulong boot_private_data)
 {
 	struct src *src_reg;
 	u32 stack, pc;
